+++ internal/ledger/ledger.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 )
 
-func CalculateBalances(chain []*block.Block) map[string]float64 {
-	balances := make(map[string]float64)
+func CalculateBalances(chain []*block.Block) map[string]uint64 {
+	balances := make(map[string]uint64)
 
 	for _, b := range chain {
 		for _, tx := range b.Transactions {
@@ -21,7 +21,7 @@ func CalculateBalances(chain []*block.Block) map[string]float64 {
 }
 
 // CalculateAvailableBalances returns the balance available to spend (ledger minus pending outbounds)
-func CalculateAvailableBalances(chain []*block.Block, pendingPool []block.Transaction) map[string]float64 {
+func CalculateAvailableBalances(chain []*block.Block, pendingPool []block.Transaction) map[string]uint64 {
 	balances := CalculateBalances(chain)
 
 	// deduct the pending pool transactions to prevent double spending
@@ -33,14 +33,14 @@ func CalculateAvailableBalances(chain []*block.Block, pendingPool []block.Transa
 	return balances
 }
 
-func ValidateTransaction(tx block.Transaction, balances map[string]float64) error {
+func ValidateTransaction(tx block.Transaction, balances map[string]uint64) error {
 	if tx.Amount <= 0 {
 		return errors.New("Amount must be Greater than 0")
 	}
 
 	if tx.Sender != "FAUCET" && tx.Sender != "COINBASE" {
 		if balances[tx.Sender] < tx.Amount {
-			return fmt.Errorf("Insufficent funds: need %f, but have %f", tx.Amount, balances[tx.Sender])
+			return fmt.Errorf("Insufficent funds: need %d, but have %d", tx.Amount, balances[tx.Sender])
 		}
 	}
 	return nil
